Build tools and libs upgrade * Converted VS2013 project files to VS2015. * WxWidgets upgraded to 3.1.0. * Lots of warnings fixed.

+++ creator/creator.cpp
@@ -16,7 +16,7 @@ void read_file(const char * filename, std::vector<char> & vec)
 	ifstream::pos_type fileSize = ifs.tellg();
 	ifs.seekg(0, ios::beg);
 
-	vec.resize(fileSize);
+	vec.resize((size_t)fileSize);
 	ifs.read(&vec[0], fileSize);
 }
